+++ src/main/java/neko/project/meusgastos/gastos/GastosRepository.java
@@ -2,5 +2,5 @@ package neko.project.meusgastos.gastos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GastosRepository extends JpaRepository<Gastos, Long> {
+public interface GastosRepository extends JpaRepository<Gastos, Integer> {
 }
